msg-block file
