+++ extension/chrome/elements/pgp_block_modules/pgp_block_render_module.ts
@@ -4,7 +4,7 @@
 
 import { PgpBlockView } from '../pgp_block.js';
 import { Xss } from '../../../js/common/platform/xss.js';
-import { MsgBlock, Mime } from '../../../js/common/core/mime.js';
+import { Mime } from '../../../js/common/core/mime.js';
 import { Att } from '../../../js/common/core/att.js';
 import { Buf } from '../../../js/common/core/buf.js';
 import { VerifyRes, PgpMsg } from '../../../js/common/core/pgp-msg.js';
@@ -12,6 +12,7 @@ import { BrowserMsg } from '../../../js/common/browser/browser-msg.js';
 import { Ui } from '../../../js/common/browser/ui.js';
 import { Store } from '../../../js/common/platform/store.js';
 import { Catch } from '../../../js/common/platform/catch.js';
+import { MsgBlock } from '../../../js/common/core/msg-block.js';
 
 export class PgpBlockViewRenderModule {
