+++ extension/js/common/xss_safe_factory.ts
@@ -7,13 +7,13 @@
 import { Catch } from './platform/catch.js';
 import { Str, Dict, UrlParams, Url } from './core/common.js';
 import { Att } from './core/att.js';
-import { MsgBlock } from './core/mime.js';
 import { Browser } from './browser/browser.js';
 import { Xss } from './platform/xss.js';
 import { SendAsAlias } from './platform/store.js';
 import { Ui } from './browser/ui.js';
 import { WebMailName } from './browser/env.js';
 import { PgpArmor } from './core/pgp-armor.js';
+import { MsgBlock, MsgBlockParser } from './core/msg-block.js';
 
 type Placement = 'settings' | 'settings_compose' | 'default' | 'dialog' | 'gmail' | 'embedded' | 'compose';
 export type WebmailVariantString = undefined | 'html' | 'standard' | 'new';
@@ -287,7 +287,7 @@ export class XssSafeFactory {
    * When edited, REQUEST A SECOND SET OF EYES TO REVIEW CHANGES
    */
   public static replaceRenderableMsgBlocks = (factory: XssSafeFactory, origText: string, msgId?: string, senderEmail?: string, isOutgoing?: boolean) => {
-    const { blocks } = PgpArmor.detectBlocks(origText);
+    const { blocks } = MsgBlockParser.detectBlocks(origText);
     if (blocks.length === 1 && blocks[0].type === 'plainText') {
       return undefined; // only has single block which is plain text - meaning
     }
